* Merge from three_thoughts_on_audio_etc branch.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@5094 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sound/MP3AudioFile.cpp
@@ -131,7 +131,7 @@ MP3AudioFile::parseHeader()
          }
     }
     
-    if (foundMP3 == false || hS.length() < bufferLength)
+    if (foundMP3 == false || (int)hS.length() < bufferLength)
     {
         std::string mess = std::string("\"")+ m_fileName +
             std::string("\" doesn't appear to be a valid MP3 file");
